+++ packages/samples/headless-ssr-commerce/app/middleware.ts
@@ -1,0 +1,10 @@
+import {NextRequest, NextResponse} from 'next/server';
+
+export default function middleware(request: NextRequest) {
+  const response = NextResponse.next();
+  const requestHeaders = new Headers(request.headers);
+  const uuid = crypto.randomUUID();
+  requestHeaders.set('x-coveo-client-id', uuid);
+  response.headers.set('x-coveo-client-id', uuid);
+  return response;
+}